add gc to db

+++ src/db/kv/btree_map.rs
@@ -15,18 +15,34 @@ use std::fmt::Debug;
 /// and then clone the [`BTreeMap`] manually and create a new via [`BTreeMapDb::from_map`].
 #[derive(Default)]
 pub struct BTreeMapDb {
+    gc_enabled: bool,
     db: BTreeMap<Box<[u8]>, Box<[u8]>>,
 }
 
 impl BTreeMapDb {
     /// Create a new empty `BTreeMapDb`.
-    pub fn new() -> Self {
-        Self::default()
+    pub fn new(gc_enabled: bool) -> Self {
+        Self {
+            gc_enabled,
+            db: BTreeMap::new(),
+        }
     }
 
     /// Create a new `BTreeMapDb` from a `BTreeMap`.
-    pub fn from_map(db: BTreeMap<Box<[u8]>, Box<[u8]>>) -> Self {
-        Self { db }
+    pub fn from_map(gc_enabled: bool, db: BTreeMap<Box<[u8]>, Box<[u8]>>) -> Self {
+        Self { gc_enabled, db }
+    }
+
+    /// Enable or disable garbage collection.
+    #[inline]
+    pub fn set_gc_enabled(&mut self, gc_enabled: bool) {
+        self.gc_enabled = gc_enabled;
+    }
+
+    /// Check if garbage collection is enabled.
+    #[inline]
+    pub fn is_gc_enabled(&self) -> bool {
+        self.gc_enabled
     }
 
     /// Get the inner `BTreeMap`.
@@ -34,7 +50,7 @@ impl BTreeMapDb {
         &self.db
     }
 
-    /// Get the inner `BTreeMap`.
+    /// Into the inner `BTreeMap`.
     pub fn into_inner(self) -> BTreeMap<Box<[u8]>, Box<[u8]>> {
         self.db
     }
@@ -60,8 +76,17 @@ impl KVDatabase for BTreeMapDb {
     fn get(&self, k: &[u8]) -> Result<Option<impl AsRef<[u8]>>, Self::Error> {
         Ok(self.db.get(k))
     }
+
+    fn gc_enabled(&self) -> bool {
+        self.gc_enabled
+    }
+
     fn remove(&mut self, k: &[u8]) -> Result<(), Self::Error> {
-        self.db.remove(k);
+        if self.gc_enabled {
+            self.db.remove(k);
+        } else {
+            warn!("garbage collection is disabled, remove is ignored");
+        }
         Ok(())
     }
