+++ src/db/kv/hash_map.rs
@@ -13,26 +13,42 @@ use std::fmt::Debug;
 /// and then clone the [`HashMapDb`] manually and create a new via [`HashMapDb::from_map`].
 #[derive(Default)]
 pub struct HashMapDb {
+    gc_enabled: bool,
     db: HashMap<Box<[u8]>, Box<[u8]>>,
 }
 
 impl HashMapDb {
-    /// Create a new empty `HashMapDb`.
-    pub fn new() -> Self {
-        Self::default()
+    /// Create a new empty [`HashMapDb`].
+    pub fn new(gc_enabled: bool) -> Self {
+        Self {
+            gc_enabled,
+            db: HashMap::new(),
+        }
     }
 
-    /// Create a new `BTreeMapDb` from a `BTreeMap`.
-    pub fn from_map(db: HashMap<Box<[u8]>, Box<[u8]>>) -> Self {
-        Self { db }
+    /// Create a new [`HashMapDb`] from a [`HashMap`](std::collections::HashMap).
+    pub fn from_map(gc_enabled: bool, db: HashMap<Box<[u8]>, Box<[u8]>>) -> Self {
+        Self { gc_enabled, db }
     }
 
-    /// Get the inner `BTreeMap`.
+    /// Enable or disable garbage collection.
+    #[inline]
+    pub fn set_gc_enabled(&mut self, gc_enabled: bool) {
+        self.gc_enabled = gc_enabled;
+    }
+
+    /// Check if garbage collection is enabled.
+    #[inline]
+    pub fn is_gc_enabled(&self) -> bool {
+        self.gc_enabled
+    }
+
+    /// Get the inner [`HashMap`](std::collections::HashMap).
     pub fn inner(&self) -> &HashMap<Box<[u8]>, Box<[u8]>> {
         &self.db
     }
 
-    /// Get the inner `BTreeMap`.
+    /// Into the inner [`HashMap`](std::collections::HashMap).
     pub fn into_inner(self) -> HashMap<Box<[u8]>, Box<[u8]>> {
         self.db
     }
@@ -59,8 +75,16 @@ impl KVDatabase for HashMapDb {
         Ok(self.db.get(k))
     }
 
+    fn gc_enabled(&self) -> bool {
+        self.gc_enabled
+    }
+
     fn remove(&mut self, k: &[u8]) -> Result<(), Self::Error> {
-        self.db.remove(k);
+        if self.gc_enabled {
+            self.db.remove(k);
+        } else {
+            warn!("garbage collection is disabled, remove is ignored");
+        }
         Ok(())
     }
